Add toggleTask function to the App component.
This change adds a toggleTask function to the App component. It takes an id and isCompleted argument. It uses the setTasks function to update the tasks state. It resulting state is the same as the previous state, except for the task with the matching id. That task's isCompleted property is set to the value of the isCompleted argument.